requirements.txt
absl-py==2.0.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.0
async-timeout==4.0.3
backcall==0.2.0
black==21.4b2
blinker==1.7.0
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cvbase==0.5.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
detectron2==0.6+cpu
easydict==1.10
etaprogress==1.1.1
exceptiongroup==1.1.3
executing==2.0.0
filelock @ file:///croot/filelock_1672387128942/work
Flask==3.0.0
fonttools==4.43.1
fsspec==2023.9.2
future==0.18.3
fvcore==0.1.5.post20221221
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
httplib2==0.22.0
hydra-core==1.3.2
icecream==2.1.3
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.5
imageio-ffmpeg==0.4.3
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.9
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
kornia==0.5.0
lazy_loader==0.3
Markdown==3.5
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.8.0
matplotlib-inline==0.1.6
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp39-cp39-linux_x86_64.whl#sha256=6c17acd1e1c4a5ebb5d14a2b7abe3d296469c0dd858a8b623abc1f4603184a9d
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==3.11.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
PyDrive2==1.17.0
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pyspng==0.1.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq==25.1.1
qudida==0.0.4
redis==5.0.1
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
service-streamer==0.1.2
shapely==2.0.1
six==1.16.0
stack-data==0.6.3
sympy @ file:///croot/sympy_1668202399572/work
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.9.26
toml==0.10.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth==0.2.8
Werkzeug==3.0.0
wldhx.yadisk-direct==0.0.6
yacs==0.1.8
zipp==3.17.0
